requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-CORS==5.0.1
PyMySQL==1.1.1
cryptography==44.0.0
bcrypt==4.2.1
marshmallow==3.23.2
python-dotenv==1.0.1
Werkzeug==3.1.3
